{{cookiecutter.repo_name}}/requirements.txt
Django==1.7.4
django-extensions==1.5.0
djangorestframework==3.0.4
six==1.9.0
wsgiref==0.1.2
